create a custom popup on deletion

+++ src/app/employees/employee-list/employee-list.component.ts
@@ -5,6 +5,7 @@ import { DepartmentService } from 'src/app/shared/department.service';
 import { MatDialog, MatDialogConfig } from '@angular/material';
 import { EmployeeComponent } from '../employee/employee.component';
 import { NotificationService } from 'src/app/shared/notification.service';
+import { DialogService } from 'src/app/shared/dialog.service';
 
 @Component({
   selector: 'app-employee-list',
@@ -30,7 +31,8 @@ export class EmployeeListComponent implements OnInit {
   constructor(private service: EmployeeService,
     private departmentService: DepartmentService,
     private dialog: MatDialog, 
-    private notificationService: NotificationService) { }
+    private notificationService: NotificationService,
+    private dialogService: DialogService) { }
 
   //in ngOnInit I get the employee entries from firebase. getEmployees returns an observable, i have to subscribe to! 
   //inside the subscribe function I have to convert the data into an array
@@ -94,9 +96,12 @@ export class EmployeeListComponent implements OnInit {
   }
 
   onDelete($key) {
-    if(confirm('Are you sure to delete this record ?')){
-    this.service.deleteEmployee($key);
-    this.notificationService.warn('! Deleted sucessfully');
+    this.dialogService.openConfirmDialog('Are you sure to delete this record ?')
+    .afterClosed().subscribe(res => {
+      if(res) {
+        this.service.deleteEmployee($key);
+        this.notificationService.warn('! Deleted sucessfully');
+      }
+    });
   }
-}
 }
